Using the pyte terminal library instead of shellinabox.
Shellinabox provides a complete and fast vt100 implementation in
javascript - it even runs and communicates well in the embedded webkit - but
changing its code to also render iframes and html elements instead of plain
chars only prooved to be to hard for me. I blame my inability to extend that
code on its C-ish nature with lots of global state and most code mixing
presentation, session handling and terminal command interpretation together.

Pyte instead is pure python and provides a nice abstraction around streams and
screen. The code is well written and it seems easy to implement a Screen class
rendering to html instead of a character matrix.